requirements.txt
Django
djangorestframework
django-cors-headers
crewai
crewai-tools
transformers
torch
sentence-transformers
huggingface-hub
langchain
langchain-community
openai
requests
pandas
numpy
scikit-learn
nltk
spacy
textblob
matplotlib
seaborn
plotly
python-dotenv
python-decouple
openpyxl
xlrd
python-magic
beautifulsoup4
reportlab
whitenoise
gunicorn
psycopg2-binary
celery
redis
channels
channels-redis
asgiref
pytest
pytest-django
factory-boy
coverage
black
flake8
django-debug-toolbar
django-extensions
cryptography
django-environ
python-jose
sentry-sdk
drf-spectacular
google-generativeai
python-dateutil
pytz
six
urllib3
certifi
django-cors-headers
djangorestframework
channels
channels-redis
daphne
